### 2018-02-01 #562 @whedon
[PRE REVIEW]: EFAshiny: An User-Friendly Shiny Application for Exploratory Factor Analysis
**Submitting author:** @PsyChiLin (<a href="http://orcid.org/0000-0002-4381-7163">Chi-Lin Yu</a>)
**Repository:** <a href="https://github.com/PsyChiLin/EFAshiny" target ="_blank">https://github.com/PsyChiLin/EFAshiny</a>
**Version:** 1.0.0
