+++ paper/paper.bib
@@ -15,7 +15,7 @@
 
 @book{ggplot2,
 	Author = {Hadley Wickham},
-	Journal = { Springer},
+	Journal = {Springer},
 	Title = {ggplot2: elegant graphics for data analysis},
 	Address = {New York},
 	Year = 2016}
@@ -27,7 +27,6 @@
   note = {R package version 1.7.8},
   url = {https://cran.r-project.org/web/packages/psych/},
 }
-
 
 @Manual{corp,
   title = {corrplot: Visualization of a Correlation Matrix},
@@ -45,36 +44,42 @@
 }
 
 @article{EGA,
-  title={Exploratory graph analysis: a new approach for estimating the number of dimensions in psychological research},
-  author={Hudson Golino, Sacha Epskamp},
-  journal={Plos One},
-  volume={12},
-  number={6},
-  pages={26},
-  year={2017},
-  publisher={Public Library of Science},
-  doi={https://doi.org/10.1371/journal.pone.0174035}
+    author = {Golino, Hudson F and Epskamp, Sacha},
+    journal = {PLOS ONE},
+    publisher = {Public Library of Science},
+    title = {Exploratory graph analysis: A new approach for estimating the number of dimensions in psychological research},
+    year = {2017},
+    month = {06},
+    volume = {12},
+    url = {https://doi.org/10.1371/journal.pone.0174035},
+    pages = {1-26},
+    number = {6},
+    doi = {10.1371/journal.pone.0174035}
 }
 
 @article{efaguide,
-  title={Use of exploratory factor analysis in published research: Common errors and some comment on improved practice},
-  author={Robin K Henson, J Kyle Roberts},
-  journal={Educational and Psychological measurement},
-  volume={66},
-  number={3},
-  pages={393-416},
-  year={2006},
-  publisher={Sage Publications},
-  doi={https://doi.org/10.1177/0013164405282485}
+author = {Robin K. Henson and J. Kyle Roberts},
+title ={Use of Exploratory Factor Analysis in Published Research: Common Errors and Some Comment on Improved Practice},
+journal = {Educational and Psychological Measurement},
+volume = {66},
+number = {3},
+pages = {393-416},
+year = {2006},
+doi = {10.1177/0013164405282485},
+URL = {https://doi.org/10.1177/0013164405282485},
+eprint = {https://doi.org/10.1177/0013164405282485}
 }
 
 @article{mclaughlin1992potential,
-  title={Potential language and attentional networks revealed through factor analysis of rCBF data measured with SPECT},
-  author={McLaughlin, Thomas and Steinberg, Bruce and Christensen, Birger and Law, Ian and Parving, Agnete and Friberg, Lars},
-  journal={Journal of Cerebral Blood Flow \& Metabolism},
-  volume={12},
-  number={4},
-  pages={535--545},
-  year={1992},
-  publisher={SAGE Publications Sage UK: London, England}
+author = {Thomas McLaughlin and Bruce Steinberg and Birger Christensen and Ian Law and Agnete Parving and Lars Friberg},
+title ={Potential Language and Attentional Networks Revealed through Factor Analysis of rCBF Data Measured with SPECT},
+journal = {Journal of Cerebral Blood Flow \& Metabolism},
+volume = {12},
+number = {4},
+pages = {535-545},
+year = {1992},
+doi = {10.1038/jcbfm.1992.77},
+note ={PMID: 1618932},
+URL = {https://doi.org/10.1038/jcbfm.1992.77},
+eprint = {https://doi.org/10.1038/jcbfm.1992.77}
 }
