### 2018-02-01 #562 @PsyChiLin
@cMadan
After searching papers in JOSS, I find a few potential reviewers who have submitted or have reviewed psychology-related projects recently.
@DominiqueMakowski @andrewheusser @RichardLitt
Also look forward to your suggestions !
