### 2018-02-06 #567 @DominiqueMakowski
@cMadan @PsyChiLin

EFAshiny is both an R package and an online app designed to provide a clear, straightforward pipeline to exploratory factor analysis, a widely used statistical method. It is a nice package, very well done, that will be very useful to many people in social science and related fields. Installation and launch work as expected, and everything looks in order within the paper and the documentation.

Below are some observations:

### General
1) The documentation on how to use the package is ok (with example data).
2) There is no CRAN depot
3) There is no github release, which will be necessary for further publication
4) There is no automated test coverage check, which could be a useful addition
5) Dependencies are stated clearly

### Specific
1) In "numeric statistics" (data summary): enable the display of Median & MAD would be useful
2) For the "distribution" tab in data summary, it would be nice to 1) have the possibility to have density plots rather than histograms and 2) to render these plots dynamically using ggplotly. I believe this wouldn't be complicated to implement.
3) Correlation plot (data summary) could possibly benefit from being generated with ggcorrplot instead of corrplot, allowing for increased flexibility and visual coherence. (altough the aesthetic gain is indeed relative).
4) Factor retention: why not allow the use of different methods for factor retention, or methods summary (see the [psycho's](https://cran.r-project.org/web/packages/psycho/vignettes/overview.html#how-many-factorscomponents-to-retain) function and [this](https://github.com/neuropsychology/psycho.R/issues/29) unsolved issue)
5) Exploratory graph analysis, It would be a small yet useful addition to be able to plot the graph with different layouts (tree, circle, ..., see [qgraph](http://sachaepskamp.com/documentation/qgraph/) documentation)
6) It would be great to be able to extract figures in R (in ggplot format or so), to further use them in rmarkdown reports or customize them.
7) In the same vein, a button to get the R code used to generate the results would be very useful for learning R and implement automated processing scripts.
8) I understand this is a package for exploraty factor analysis (hence its name), however, I'm curious on the possibility of adding a CFA tab (at the end, refitting the model using CFA (and, maybe at the beginning, an option to split the data into a training and a test dataset))... Your tool would be then sufficient for the entire, end-to-end, factor analysis.
9) In general, the feature that lacks the most is the easy possibility to interface your app with a programming R usage. In other words, the possibility to use your package within a more general script pipeline. That includes the possibility to extract the data (for example, the loadings with their confidence intervals), the code used at each step to generate the results etc., to be able to further reproduce the entire analysis without launching the app and clicking on the buttons.

These are minor suggestions that do not hinder the general quality of the package. Therefore, I'm very positive about its publication.

Best,
