### 2018-02-06 PsyChiLin/EFAshiny#1 @PsyChiLin
JOSS Review - General
General points from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/567):

1. The documentation on how to use the package is ok (with example data).
2. There is no CRAN depot
3. There is no github release, which will be necessary for further publication
4. There is no automated test coverage check, which could be a useful addition
5. Dependencies are stated clearly
