### 2018-02-06 PsyChiLin/EFAshiny#2 @PsyChiLin
JOSS Review - Specific 1 - Numeric statistics
Specific point 1 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/567):

In "numeric statistics" (data summary): enabling the display of Median & MAD would be useful
