### 2018-02-06 PsyChiLin/EFAshiny#3 @PsyChiLin
JOSS Review - Specific 2 - Distribution
Specific point 2 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/567):

For the "distribution" tab in data summary, it would be nice to 1) have the possibility to have density plots rather than histograms and 2) to render these plots dynamically using ggplotly. I believe this wouldn't be complicated to implement.
