### 2018-02-06 PsyChiLin/EFAshiny#4 @PsyChiLin
JOSS Review - Specific 3 - Correlation plot
Specific point 3 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/567):

Correlation plot (data summary) could possibly benefit from being generated with ggcorrplot instead of corrplot, allowing for increased flexibility and visual coherence. (altough the aesthetic gain is indeed relative).
