### 2018-02-06 PsyChiLin/EFAshiny#5 @PsyChiLin
JOSS Review - Specific 4 - Factor retention
Specific point 4 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/567):

Factor retention: why not allow the use of different methods for factor retention, or methods summary (see the [psycho's](https://cran.r-project.org/web/packages/psycho/vignettes/overview.html#how-many-factorscomponents-to-retain) function and [this](https://github.com/neuropsychology/psycho.R/issues/29) unsolved issue)
