### 2018-02-06 PsyChiLin/EFAshiny#6 @PsyChiLin
JOSS Review - Specific 5 - Exploratory graph analysis
Specific points 5 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/567):

Exploratory graph analysis, It would be a small yet useful addition to be able to plot the graph with different layouts (tree, circle, ..., see [qgraph](http://sachaepskamp.com/documentation/qgraph/) documentation)
