### 2018-02-06 PsyChiLin/EFAshiny#7 @PsyChiLin
JOSS Review - Specific 6, 7, 9 - R programming usage
Specific point 6, 7, 9 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/567):

6. It would be great to be able to extract figures in R (in ggplot format or so), to further use them in rmarkdown reports or customize them.

7. In the same vein, a button to get the R code used to generate the results would be very useful for learning R and implement automated processing scripts.

9. In general, the feature that lacks the most is the easy possibility to interface your app with a programming R usage. In other words, the possibility to use your package within a more general script pipeline. That includes the possibility to extract the data (for example, the loadings with their confidence intervals), the code used at each step to generate the results etc., to be able to further reproduce the entire analysis without launching the app and clicking on the buttons.
