### 2018-02-06 PsyChiLin/EFAshiny#8 @PsyChiLin
JOSS Review - Specific 8 - CFA
Specific point 8 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/567):

I understand this is a package for exploraty factor analysis (hence its name), however, I'm curious on the possibility of adding a CFA tab (at the end, refitting the model using CFA (and, maybe at the beginning, an option to split the data into a training and a test dataset))... Your tool would be then sufficient for the entire, end-to-end, factor analysis.
