### 2018-02-06 #567 @PsyChiLin
@DominiqueMakowski Thanks for these useful suggestions!
I have opened several issues for all of them to track their progress. Will improve this project in the next couple of days.
@cMadan Will let you know the progress after I have advanced this submission.
Thanks again !
