### 2018-02-07 f1ed67a
1.0.0 - median & MAD
