+++ inst/efas/functions/my_summary.R
@@ -1,4 +1,4 @@
 my_summary <- function(x) {
-        funs <- c(mean, sd, skewness, kurtosis)
+        funs <- c(mean, sd, skewness, kurtosis, median, mad)
         sapply(funs, function(f)f(x, na.rm = TRUE))
 }
