### 2018-02-07 PsyChiLin/EFAshiny#2 @PsyChiLin
I've added Median and MAD at the default display.
