### 2018-02-07 PsyChiLin/EFAshiny#3 @PsyChiLin
I've added the possibility to enable the plotting of density plots.
The histograms are still there, because they can be somehow useful for item-level datasets.
Also render these two plots using plotly.
Dynamic displays are really suitable for these two plots ! Thanks for the suggestions haha :))
