### 2018-02-07 PsyChiLin/EFAshiny#4 @PsyChiLin
As suggested, I have added another tab for ggcorplot. Hope it can increase our visual coherence.
The original corplot was not deleted because it can still be a useful visualization.
By the way, our design logic is to allow users to easily implement EFA without additional worries; thus, minor options (i.e., color, label, order...etc) are set with reasonable defaults. Users do not need to worry about lots of arguments in the analyses :))
