### 2018-02-07 PsyChiLin/EFAshiny#6 @PsyChiLin
Have added the layout options for EGA plot (Layout "spring" is still the default as suggested in original EGA function).
