### 2018-02-07 PsyChiLin/EFAshiny#5 @PsyChiLin
I have provided the possibility to estimate additional indices (e.g., VSS1 & VSS2).
Also really thanks for the useful suggestion for methods summary. I have added a summary table for several provided methods, allowing users to easily make a decision. Good reference and great idea in the 'psycho' reference !
