+++ paper/paper.bib
@@ -28,12 +28,15 @@
   url = {https://cran.r-project.org/web/packages/psych/},
 }
 
-@Manual{corp,
-  title = {corrplot: Visualization of a Correlation Matrix},
-  author = {Taiyun Wei},
-  year = {2017},
-  note = {R package version 0.84},
-  url = {https://cran.r-project.org/web/packages/corrplot/},
+@article{psycho,
+        author = {Dominique Makowski},
+        title = {The psycho Package: an Efficient and Publishing-Oriented Workflow for Psychological Science},
+        journal = {Journal of Open Source Software},
+        year = {2018},
+        volume = {3},
+        number = {22},
+        pages = {470},
+        doi = {10.21105/joss.00470}
 }
 
 @article{r,
