+++ inst/efas/functions/faplot.R
@@ -5,503 +5,6 @@ faplot <- function(data,n.obs = NULL,quant = 0.95,fm = "promax",n.iter = 200){
         library(ggplot2)
         library(psych)
         #Sys.sleep(2)
-        fa.parallel <- function (x, n.obs = NULL, fm = "minres", fa = "both", main = "Parallel Analysis Scree Plots", 
-                                 n.iter = 20, error.bars = FALSE, se.bars = FALSE, SMC = FALSE, 
-                                 ylabel = NULL, show.legend = TRUE, sim = TRUE, quant = 0.95, 
-                                 cor = "cor", use = "pairwise", plot = TRUE, correct = 0.5) {
-                cl <- match.call()
-                ci <- 1.96
-                arrow.len <- 0.05
-                nsub <- dim(x)[1]
-                nvariables <- dim(x)[2]
-                resample <- TRUE
-                if ((isCorrelation(x)) && !sim) {
-                        warning("You specified a correlation matrix, but asked to just resample (sim was set to FALSE).  This is impossible, so sim is set to TRUE")
-                        sim <- TRUE
-                        resample <- FALSE
-                }
-                if (!is.null(n.obs)) {
-                        nsub <- n.obs
-                        rx <- x
-                        resample <- FALSE
-                        if (dim(x)[1] != dim(x)[2]) {
-                                warning("You specified the number of subjects, implying a correlation matrix, but do not have a correlation matrix, correlations found ")
-                                switch(cor, cor = {
-                                        rx <- cor(x, use = use)
-                                }, cov = {
-                                        rx <- cov(x, use = use)
-                                        covar <- TRUE
-                                }, tet = {
-                                        rx <- tetrachoric(x, correct = correct)$rho
-                                }, poly = {
-                                        rx <- polychoric(x, correct = correct)$rho
-                                }, mixed = {
-                                        rx <- mixedCor(x, use = use, correct = correct)$rho
-                                }, Yuleb = {
-                                        rx <- YuleCor(x, , bonett = TRUE)$rho
-                                }, YuleQ = {
-                                        rx <- YuleCor(x, 1)$rho
-                                }, YuleY = {
-                                        rx <- YuleCor(x, 0.5)$rho
-                                })
-                                if (!sim) {
-                                        warning("You specified a correlation matrix, but asked to just resample (sim was set to FALSE).  This is impossible, so sim is set to TRUE")
-                                        sim <- TRUE
-                                        resample <- FALSE
-                                }
-                        }
-                }
-                else {
-                        if (isCorrelation(x)) {
-                                warning("It seems as if you are using a correlation matrix, but have not specified the number of cases. The number of subjects is arbitrarily set to be 100  ")
-                                rx <- x
-                                nsub = 100
-                                n.obs = 100
-                                resample <- FALSE
-                        }
-                        else {
-                                switch(cor, cor = {
-                                        rx <- cor(x, use = use)
-                                }, cov = {
-                                        rx <- cov(x, use = use)
-                                        covar <- TRUE
-                                }, tet = {
-                                        rx <- tetrachoric(x, correct = correct)$rho
-                                }, poly = {
-                                        rx <- polychoric(x, correct = correct)$rho
-                                }, mixed = {
-                                        rx <- mixedCor(x, use = use, correct = correct)$rho
-                                }, Yuleb = {
-                                        rx <- YuleCor(x, , bonett = TRUE)$rho
-                                }, YuleQ = {
-                                        rx <- YuleCor(x, 1)$rho
-                                }, YuleY = {
-                                        rx <- YuleCor(x, 0.5)$rho
-                                })
-                        }
-                }
[…]
