### 2018-02-08 93820fc
1.0.0 - Editor Mode, R markdown
