### 2018-02-08 0b29079
1.0.0 - Editor Mode
