+++ inst/efas/ui.R
@@ -16,22 +16,14 @@ if(!require(shinyAce)) {require(shinyAce)}
 if(!require(RCurl)) {require(RCurl)}
 
 code = '
+### Overview
+This tutorial aims to demonstrate how to perform the fundamental analyses in [`EFAshiny`](https://psychilin.shinyapps.io/EFAshiny/) using R code. By following each line in this tutorial, you can learn basic R programming and can implement automated processing script in future analyses. Also, you can further use the code in `EFAshiny` within a script pupeline without launching the app. However, we still suggest you to use [`EFAshiny`](https://psychilin.shinyapps.io/EFAshiny/) APP :)) You can also have a look at our [tutorial](https://github.com/PsyChiLin/EFAshiny) of `EFAshiny` GUI to make an easy comparison while writing the code. Have fun with [`EFAshiny`](https://psychilin.shinyapps.io/EFAshiny/) and `R` !
+
 ### Preparation
-```{r, message=FALSE, warning=FALSE, echo = F}
-if(!require(shiny)) {require(shiny)}
-if(!require(shinythemes)) {require(shinythemes)}
-if(!require(ggplot2)) {require(ggplot2)}
-if(!require(psych)) {require(psych)}
-if(!require(EFAutilities)) {require(EFAutilities)}
-if(!require(corrplot)) {require(corrplot)}
-if(!require(reshape2)) {require(reshape2)}
-if(!require(moments)) {require(moments)}
-if(!require(gridExtra)) {require(gridExtra)}
-if(!require(qgraph)) {require(qgraph)}
-if(!require(bootnet)) {require(bootnet)}
-if(!require(igraph)) {require(igraph)}
-if(!require(ggcorrplot)) {require(ggcorrplot)}
-if(!require(RCurl)) {require(RCurl)}
+First of all, we should load all the packages and fucntions in `EFAshiny` into `R` session. If you do not have these package installed, please use `install.package("packagename")` to perfrom installations.
+```{r, message=FALSE, warning=FALSE}
+require(ggplot2);require(psych);require(corrplot);require(reshape2);require(moments);require(gridExtra)
+require(qgraph);require(bootnet);require(igraph);require(ggcorrplot);require(RCurl)
 source("https://raw.githubusercontent.com/PsyChiLin/EFAshiny/master/inst/efas/functions/my_summary.R")
 source("https://raw.githubusercontent.com/PsyChiLin/EFAshiny/master/inst/efas/functions/faplot.R")
 source("https://raw.githubusercontent.com/PsyChiLin/EFAshiny/master/inst/efas/functions/bootEGA.R")
@@ -42,13 +34,14 @@ source("https://raw.githubusercontent.com/PsyChiLin/EFAshiny/master/inst/efas/fu
 ```
 
 ### Data Input
-
+Read the demonsration data [`RSE`](https://github.com/PsyChiLin/EFAshiny/blob/master/RSE/RSE.csv) into `R`. We assign it to a object called `dta`. Then using `head` to observe and explore this data. See our [tutorial](https://github.com/PsyChiLin/EFAshiny) for data description.
 ```{r}
 dta <- read.csv(text=getURL("https://raw.githubusercontent.com/PsyChiLin/EFAshiny/master/RSE/RSE.csv"))
 head(dta)
 ```
 
 ### Data Summary
+Several kinds of data summary can be perform in `EFAshiny`. We demonstrate them using basic code. The package `ggplot2` is a useful package for plotting. We use it to plot the histograms, the density plots, and the correlation matrix. You can also check the documentation of [`ggplot2`](https://cran.r-project.org/web/packages/ggplot2/ggplot2.pdf).
 
 ##### Numeric Statistic
 ```{r}
[…]
