### 2018-02-08 26b6e87
1.0.0 - Editor
