### 2018-02-08 7a889a6
1.0.0 - Editor
