### 2018-02-08 PsyChiLin/EFAshiny#8 @PsyChiLin
Adding a CFA tab maybe a cool idea. But, in our view, it is not the major focus for `EFAshiny`. We aim to provide a tool for EFA and exploratory data analysis rather than testing. Although we do not directly provide the CFA options in the tool, users can still perform CFA using `Editor` tab (Github version). The tutorial and the README file also update this related information. Thanks for the comments !
