### 2018-02-08 PsyChiLin/EFAshiny#7 @PsyChiLin
Thanks for these comments. I have added a `Editor` tab with several quick examples to allow users to learn how to perform analyses or produce plots similar to those in `EFAshiny` GUI. I think a button to get R code is not necessary, because users can use the `showcase` function in `shiny` to understand the underlying code. Instead, they need some useful and easy examples to learn R and further automate their analyses in the near future. Thus, I have designed a `Editor` tab for users who use the Github version of `EFAshiny`. They can adopt `EFAshiny` within a more general script framework as you suggested right now haha.
In summary, users who want to further understand EFAshiny or learn R can (1) see the code in `Editor` tab of `EFAshiny` GUI, (2) download the [R markdown file](https://github.com/PsyChiLin/EFAshiny/blob/master/EFAshiny_RCodeDemo.Rmd) similar to the code in editor mode, (3) see the same R markdown file in [this public link](http://rpubs.com/Chi-Lin/EFAshiny_R_Code_Demo), (4) use `showcase` function in `shiny`, and (5) directly see the code in `server.R` and `ui.R`.
Have also updated information in the paper and the README file.
Thanks for useful suggestions !
