+++ README.Rmd
@@ -45,11 +45,6 @@ install.packages("devtools")
 install.packages("shiny")
 ```
 
-Directly run `EFAshiny`:
-```{r, eval=FALSE}
-runGitHub("EFAshiny", "PsyChiLin")
-```
-
 Install and launch `EFAshiny`:
 ```{r, eval=FALSE}
 devtools::install_github("PsyChiLin/EFAshiny")
