### 2018-02-08 0eafe29
1.0.0 - Editor
