+++ inst/efas/ui.R
@@ -15,6 +15,7 @@ if(!require(ggcorrplot)) {require(ggcorrplot)}
 if(!require(shinyAce)) {require(shinyAce)}
 if(!require(RCurl)) {require(RCurl)}
 if(!require(knitr)) {require(knitr)}
+if(!require(shinydashboard)) {require(shinydashboard)}
 
 code = '
 ### Overview
