+++ README.Rmd
@@ -52,7 +52,7 @@ EFAshiny::EFAshiny()
 ```
 
 #### 2. Shiny APP version (Standard version)
-If you want to use the standard version of `EFAshiny`, installation is not required. The application is deployed on **shinyapps.io** server. The online APP version does not provide the `Editor` mode due to the stablility of shinyapp.io server. All functions are provided. <br />
+If you want to use the standard version of `EFAshiny`, installation is not required. The application is deployed on **shinyapps.io** server. This standard version has all the function except for the `Editor` tab (which is only useful for users who want to code online). Users can easily explore and analyze their data with this online APP without worrying about installation. <br />
 Have fun with `EFAshiny` : https://psychilin.shinyapps.io/EFAshiny/
 
 
@@ -184,7 +184,7 @@ In addition to the GUI, we also provide an `Editor` tab with several code demons
 
 Note that this feature can also allow the use of `lavaan` R package to perform confirmatory factor analysis (CFA), which is also a widely used method but not the main focus of `EFAshiny`. Simply input `require(lavaan)` should work (see [lavaan package](http://lavaan.ugent.be/) for details). Another useful tool is the `showcase` version of `shiny` when running the APP ( definitely, you can directly see the code in `server.R` and `ui.R`).
 
-In summary, Users who want to further understand `EFAshiny` or learn `R` can (1) see the code in `Editor` tab of `EFAshiny` GUI (as shown in figure), (2) download the R markdown file similar to the code in editor mode [here](https://github.com/PsyChiLin/EFAshiny/blob/master/EFAshiny_RCodeDemo.Rmd), (3) see the same R markdown file in [this public link](http://rpubs.com/Chi-Lin/EFAshiny_R_Code_Demo), (4) use `showcase` function in `shiny`, and (5) directly see the code in `server.R` and `ui.R`.
+In summary, Users who want to further understand `EFAshiny` or learn `R` can (1) see the code in `Editor` tab of github version `EFAshiny` GUI (as shown in figure), (2) download the R markdown file similar to the code in editor mode [here](https://github.com/PsyChiLin/EFAshiny/blob/master/EFAshiny_RCodeDemo.Rmd), (3) see the same R markdown file in [this public link](http://rpubs.com/Chi-Lin/EFAshiny_R_Code_Demo), (4) use `showcase` function in `shiny`, and (5) directly see the code in `server.R` and `ui.R`.
 
 ![Editor](rmdfigs/Editor.png)
 
