+++ tests/testthat/test-efashiny.R
@@ -1,0 +1,7 @@
+context("EFAshiny.R")
+
+test_that("EFAshiny work", {
+        #app <- system.file("efas",package = 'EFAshiny')
+        value <- class(EFAshiny(run = F))
+        testthat::expect_equal(value, "character")
+})
