+++ README.Rmd
@@ -14,6 +14,7 @@ knitr::opts_chunk$set(collapse=TRUE,
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 [![Build Status](https://travis-ci.org/PsyChiLin/EFAshiny.svg?branch=master)](https://travis-ci.org/PsyChiLin/EFAshiny)
+[![codecov](https://codecov.io/gh/PsyChiLin/EFAshiny/branch/master/graph/badge.svg)](https://codecov.io/gh/PsyChiLin/EFAshiny)
 [![APP Build](https://ci.appveyor.com/api/projects/status/ud2tblafomnka8hm?svg=true)](https://ci.appveyor.com/project/PsyChiLin/efashiny)
 [![Dependency Status](https://dependencyci.com/github/PsyChiLin/EFAshiny/badge)](https://dependencyci.com/github/PsyChiLin/EFAshiny)
 
