+++ .travis.yml
@@ -1,7 +1,7 @@
 language: r
 
-r_packages:
-  - covr
+r_github_packages:
+  - r-lib/covr
 
 after_success:
-  - Rscript -e 'library(covr); codecov()'
+  - Rscript -e 'covr::codecov()'
