### 2018-02-10 PsyChiLin/EFAshiny#2 @PsyChiLin
This issue has been addressed in the latest release.
