### 2018-02-10 PsyChiLin/EFAshiny#3 @PsyChiLin
This issue has been addressed in the latest release (although a little bit slow for running plotly on shinyapp.io server).
