### 2018-02-10 PsyChiLin/EFAshiny#4 @PsyChiLin
Has been addressed in the latest release.
