### 2018-02-10 PsyChiLin/EFAshiny#5 @PsyChiLin
Were addressed in the latest release.
