### 2018-02-10 PsyChiLin/EFAshiny#6 @PsyChiLin
This issue has been addressed in the latest release.
