### 2018-02-10 PsyChiLin/EFAshiny#7 @PsyChiLin
As shown in the reply, we have addressed this issues in latest version. All information are provided in the tutorial file for users to implement these tools.
