### 2018-02-10 PsyChiLin/EFAshiny#1 @PsyChiLin
Have updated the tool (`EFAshiny`), the tutorial, the README file in the latest Github release, which is good enough for us to share and control our tools. The new citations/references/dependencies  have also been updated in the paper and tutorial. Thanks for all the suggestions !
