+++ vignettes/EFAshiny.Rmd
@@ -10,38 +10,41 @@ vignette: >
 ---
 
 ## A Tutorial to EFAshiny
-[`EFAshiny`](https://psychilin.shinyapps.io/EFAshiny/) is an user-friendly application for exploratory factor analysis (EFA; Bartholomew, Knott, & Moustaki, 2011). The graphical user interface in shiny (Chang, Cheng, Allaire, Xie, & McPherson, 2017) is designed to free users from scripting in R by wrapping together various packages for data management, factor analysis, and graphics. Easy-to-follow analysis flow and reasonable default settings avoiding common errors (Henson & Roberts, 2006) are provided. Results of analysis in tables and graphs are presented on-line and can be exported.
+`EFAshiny` is an user-friendly application for exploratory factor analysis (EFA; Bartholomew, Knott, & Moustaki, 2011). The graphical user interface in shiny (Chang, Cheng, Allaire, Xie, & McPherson, 2017) is designed to free users from scripting in R by wrapping together various packages for data management, factor analysis, and graphics. Easy-to-follow analysis flow and reasonable default settings avoiding common errors (Henson & Roberts, 2006) are provided. Results of analysis in tables and graphs are presented on-line and can be exported.
 
 Key features include:
 
-- Easy-to-follow analysis flow 
-- Quick data summary
-- Graphical and numerical retention methods
-- Lots of extraction and rotation methods
-- Confidence intervals of factor loadings
-- Visualizations of factor loadings
-- Default arguments suggested by current research 
+- An easy-to-use GUI to free users from scripting in R
+- A step by step analysis flow to perform EFA 
+- Quick ways to summarize data by tables or graphs
+- Several ways to explore factor retention numerically or graphically
+- Several ways to explore factor extraction and rotation numerically or graphically
+- A display of confidence intervals for factor loadings
+- Several ways to link visualization of correlation matrix with factor structure
+- Default options are chosen according to recommendations in the literature
+- A demonstration using a real psychological scale dataset 
 
-The intended user of [`EFAshiny`](https://psychilin.shinyapps.io/EFAshiny/) is a behavioral science researcher who seeks an easy way to analyze and visualize item-level scales or datasets with a set of associated variables using EFA.
+The `EFAshiny` application is primarily aimed at behavioral researchers who want to perform EFA on a set of associated variables (e.g., item-level scale dataset). Note that it can also be used to explore FA-based connectivity analyses (McLaughlin et al., 1992) in instrument data, such as event related potentials (ERPs) and functional near-infrared spectroscopy (fNIRS). Though the major focus of `EFAshiny` is to perform EFA, it is worth noting that confirmatory factor analysis (CFA) is an useful future direction for `shiny` APP.
 
 ## Getting Started
 
-#### 1. Shiny APP version (recommended)
-If you want to use the `EFAshiny`, installation is not required. The application is deployed on **shinyapps.io** server. <br />
-Have fun with `EFAshiny` : https://psychilin.shinyapps.io/EFAshiny/
+#### 1. Github version (Full version)
+To run `EFAshiny` on your R, [`devtools`](https://cran.r-project.org/web/packages/devtools/devtools.pdf) and [`shiny`](https://cran.r-project.org/web/packages/shiny/shiny.pdf) are required. <br />
 
-#### 2. Github version
-To run `EFAshiny` on your R locally, you have to intsall [`devtools`](https://cran.r-project.org/web/packages/devtools/devtools.pdf). <br />
-Then, intsall `EFAshiny` from Github.
[…]
