### 2018-02-10 #567 @PsyChiLin
@cMadan I have tackled all the points except for the CFA one, which is not directly addressed. The latest version is 1.0.1 with github release.  All the information/references in the tutorial and paper are updated (thus I re-generate the pdf file) according to the useful suggestions from the review process. Again, many thanks for all these great suggestions, @DominiqueMakowski  !!
