### 2018-02-11 #567 @DominiqueMakowski
@cMadan @PsyChiLin I'm happy with these changes 🎉
