### 2018-02-12 #567 @cMadan
@PsyChiLin, I think we're almost all set then! Can you archive the current code on Zenodo or FigShare and tell me the DOI?
