### 2018-02-12 #567 @arfon
@PsyChiLin - your submission is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00567 ⚡️ 🚀 💥
